Move rerooting into tree_ops module
