+++ sc2ts/tree_ops.py
@@ -5,6 +5,7 @@ Miscellanous tree operations we need for sc2ts inference.
 import collections
 import logging
 import dataclasses
+from typing import List
 
 import tskit
 import numpy as np
@@ -15,6 +16,99 @@ import biotite.sequence.phylo as bsp
 from . import core
 
 logger = logging.getLogger(__name__)
+
+
+@dataclasses.dataclass
+class QuintuplyLinkedTree:
+    parent: List
+    left_child: List
+    right_child: List
+    left_sib: List
+    right_sib: List
+
+    def __str__(self):
+        s = "id\tparent\tlchild\trchild\tlsib\trsib\n"
+        for j in range(len(self.parent)):
+            s += (
+                f"{j}\t{self.parent[j]}\t"
+                f"{self.left_child[j]}\t{self.right_child[j]}\t"
+                f"{self.left_sib[j]}\t{self.right_sib[j]}\n"
+            )
+        return s
+
+    def remove_branch(self, p, c):
+        lsib = self.left_sib[c]
+        rsib = self.right_sib[c]
+        if lsib == -1:
+            self.left_child[p] = rsib
+        else:
+            self.right_sib[lsib] = rsib
+        if rsib == -1:
+            self.right_child[p] = lsib
+        else:
+            self.left_sib[rsib] = lsib
+        self.parent[c] = -1
+        self.left_sib[c] = -1
+        self.right_sib[c] = -1
+
+    def insert_branch(self, p, c):
+        assert self.parent[c] == -1, "contradictory edges"
+        self.parent[c] = p
+        u = self.right_child[p]
+        if u == -1:
+            self.left_child[p] = c
+            self.left_sib[c] = -1
+            self.right_sib[c] = -1
+        else:
+            self.right_sib[u] = c
+            self.left_sib[c] = u
+            self.right_sib[c] = -1
+        self.right_child[p] = c
+
+    def push_up(self, u):
+        """
+        Push the node u one level up the tree
+        """
+        parent = self.parent[u]
+        assert parent != -1
+        self.remove_branch(parent, u)
+        grandparent = self.parent[parent]
+        if grandparent != -1:
+            self.remove_branch(grandparent, parent)
+            self.insert_branch(grandparent, u)
+        self.insert_branch(u, parent)
+
+
+def reroot(ts, new_root, scale_time=False):
+    """
+    Reroot the tree around the specified node, keeping node IDs
+    the same.
+    """
+    assert ts.num_trees == 1
+    tree = ts.first()
+    qlt = QuintuplyLinkedTree(
+        left_child=tree.left_child_array.copy(),
+        left_sib=tree.left_sib_array.copy(),
+        right_child=tree.right_child_array.copy(),
+        right_sib=tree.right_sib_array.copy(),
+        parent=tree.parent_array.copy(),
+    )
+
+    # print()
+    # print(qlt)
+    while qlt.parent[new_root] != -1:
+        qlt.push_up(new_root)
+    # print()
+    # print(qlt)
+    tables = ts.dump_tables()
+    tables.edges.clear()
+    # NOTE: could be done with numpy so this will work for large trees.
+    for u in range(ts.num_nodes):
+        if qlt.parent[u] != -1:
+            tables.edges.add_row(0, ts.sequence_length, qlt.parent[u], u)
+    set_tree_time(tables, unit_scale=scale_time)
+    tables.sort()
+    return tables.tree_sequence()
 
 
 def biotite_to_tskit(tree):
